+++ app/GX_integration/workflows/workflow_invocation.py
@@ -17,7 +17,7 @@ from bioblend.galaxy.objects.wrappers import Workflow, Invocation, History, Data
 
 from app.galaxy import GalaxyClient
 from app.api.socket_manager import SocketManager
-from app.api.socket_enums import SocketMessageEvent, SocketMessageType
+from app.api.enums import SocketMessageEvent, SocketMessageType
 from app.enumerations import NumericLimits, InvocationTracking, JobState
 
 class WorkflowInvocationHandler:
