+++ app/GX_integration/workflows/worklfow_installer.py
@@ -13,7 +13,7 @@ from bioblend.galaxy.toolshed import ToolShedClient
 
 from app.galaxy import GalaxyClient
 from app.api.socket_manager import SocketManager
-from app.api.socket_enums import SocketMessageEvent, SocketMessageType
+from app.api.enums import SocketMessageEvent, SocketMessageType
 from app.enumerations import NumericLimits
 
 class WorkflowInstaller:
