+++ app/main.py
@@ -27,7 +27,7 @@ from app.log_setup import configure_logging
 from app.api.middleware import JWTGalaxyKeyMiddleware, RateLimiterMiddleware, DomainCORSMiddleware
 from app.api.api import api_router 
 from app.api.socket_manager import ws_manager
-from app.api.socket_enums import SocketMessageEvent
+from app.api.enums import SocketMessageEvent
 from app.orchestration.invocation_cache import InvocationCache
 from app.orchestration.invocation_tasks import InvocationBackgroundTasks
